requirements.txt
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.4
django-appconf==1.0.1
django-compressor==2.0
django-mptt==0.8.3
djangorestframework==3.3.2
drf-nested-routers==0.11.1
gunicorn==19.4.5
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
static3==0.6.1
